Publish only an ESM build Update to latest libraries Handle newest antlr4 which is now published as ESM module Fix eslint errors use npm shrinkwrap to handle package-lock issues
