--- jest.config.js
@@ -1,1 +1,0 @@
-process.env.TZ = 'UTC';
