+++ package.json
@@ -2,25 +2,21 @@
   "name": "@adobe/json-formula",
   "version": "0.1.44",
   "description": "JSONFormula Grammar and implementation",
-  "#browser": "The browser entry point is for a web browser.",
-  "browser": "lib/index.js",
-  "#main": "The main entry point is for a NodeJS-CommonJS context.",
-  "main": "lib/index.js",
-  "#module": "The module entry point is for an ES6 Modules context with import and export.",
-  "module": "dist/mjs/json-formula.js",
+  "main": "lib/json-formula.js",
+  "type": "module",
   "files": [
     "lib"
   ],
-  "devEngines": {
+  "engines": {
     "node": ">=16.0.0"
   },
   "scripts": {
     "eslint": "eslint src",
     "test": "jest",
-    "build": "rm -rf ./dist/ && rm -rf ./lib && webpack && webpack --config webpack.tutorial.babel.js && npm run docs",
-    "start": "webpack serve --config webpack-devserver.config.babel.js",
+    "build": "rm -rf ./dist/ && rm -rf ./lib && webpack && webpack --config webpack.tutorial.js && npm run docs",
+    "start": "webpack serve --config webpack-devserver.config.js",
     "docs": "jsdoc2md -f src/jmespath/*unctions.js -g none > dist/functions.md && npm run jsdocs",
-    "jsdocs" : "jsdoc2md -f src/json-formula.js -g none > build/Readme.md && jsdoc -c jsdoc.json -R build/Readme.md"
+    "jsdocs": "jsdoc2md -f src/json-formula.js -g none > build/Readme.md && jsdoc -c jsdoc.json -R build/Readme.md"
   },
   "repository": {
     "type": "git",
@@ -29,31 +25,28 @@
   "author": "jbrinkma",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@babel/core": "^7.16.10",
-    "@babel/node": "^7.16.8",
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
-    "@babel/preset-env": "^7.16.11",
-    "antlr4": "4.9.2",
-    "babel-loader": "^8.2.3",
+    "@babel/core": "^7.20.2",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/node": "^7.20.2",
+    "@babel/preset-env": "^7.20.2",
+    "antlr4": "4.11.0",
+    "babel-loader": "^9.1.0",
     "better-docs": "^2.7.2",
-    "copy-webpack-plugin": "^10.2.1",
+    "copy-webpack-plugin": "^11.0.0",
     "ebnf-parser": "^0.1.10",
-    "eslint": "^8.7.0",
+    "eslint": "^8.28.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^25.7.0",
-    "FileList": "^0.10.2",
-    "html-inline-script-webpack-plugin": "^2.0.3",
+    "eslint-plugin-jest": "^27.1.5",
+    "html-inline-script-webpack-plugin": "^3.1.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.4.7",
+    "jest": "^29.3.1",
+    "jest-cli": "^27.5.1",
     "jsdoc": "^3.6.11",
     "jsdoc-to-markdown": "^7.1.1",
-    "json-loader": "^0.5.7",
-    "raw-loader": "^4.0.2",
-    "webpack": "^5.66.0",
-    "webpack-bundle-analyzer": "^4.6.1",
-    "webpack-cli": "^4.9.1",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^5.0.0",
     "webpack-dev-server": "^4.7.3"
   }
 }
