+++ webpack.tutorial.js
@@ -11,45 +11,28 @@ governing permissions and limitations under the License.
 */
 import path from 'path';
 // eslint-disable-next-line import/no-extraneous-dependencies
+import CopyPlugin from 'copy-webpack-plugin';
+// eslint-disable-next-line import/no-extraneous-dependencies
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 
-const dist = path.resolve('.', 'dist');
+const DIST = path.resolve('.', 'dist');
 
-const defn = {
+export default {
   mode: 'production',
   entry: {
     tutorial: './src/tutorial.js',
   },
   devtool: 'source-map',
-  optimization: {
-    minimize: false,
-  },
-  devServer: {
-    port: 8085,
-    static: {
-      directory: __dirname,
-    },
-  },
-  resolve: { fallback: { fs: false } },
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: /\.(js)$/,
         exclude: /node_modules/,
         use: {
           loader: 'babel-loader',
-          options: {
-            presets: [
-              '@babel/preset-env',
-              {
-                plugins: [
-                  ['@babel/plugin-proposal-class-properties', { loose: true }],
-                  ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
-                  ['@babel/plugin-proposal-private-methods', { loose: true }],
-                ],
-              },
-            ],
-          },
+        },
+        resolve: {
+          fullySpecified: false,
         },
       },
       {
@@ -69,21 +52,19 @@ const defn = {
     ],
   },
   output: {
-    path: dist,
+    path: DIST,
     filename: '[name].js',
     libraryTarget: 'var',
     library: 'JSONFormula',
   },
   plugins: [
+    new CopyPlugin({
+      patterns: [
+        { from: './doc', to: path.resolve(DIST, 'doc') },
+      ],
+    }),
     new HtmlWebpackPlugin({
       template: 'src/index.html',
     }),
   ],
-  performance: {
-    hints: false,
-    maxEntrypointSize: 512000,
-    maxAssetSize: 512000,
-  },
 };
-
-export default defn;
